requirements.txt
# requirements.txt
requests
influxdb-client
python-dateutil
pytz
